refactor: Extract lottery data fetching into useLotteryData hook
The data fetching and color calculation logic for lottery results was duplicated in `App.jsx` and `LotteryResultsPage.jsx`. This made the code harder to maintain and prone to errors.

This change extracts this logic into a new custom React hook, `useLotteryData`. This hook now manages the state for results, loading status, and errors, and also provides the `getNumberColorClass` utility function.

The hook is parameterized to accept an `apiPrefix` to handle the different API endpoints used by `App.jsx` (`/`) and `LotteryResultsPage.jsx` (`/api`).

This refactoring improves code maintainability, reduces duplication, and makes the components simpler and easier to understand.
